move channelIdImpl class to common package as record
