--- io.openems.edge.battery.soltaro/src/io/openems/edge/battery/soltaro/common/ChannelIdImpl.java
@@ -1,26 +1,0 @@
-package io.openems.edge.battery.soltaro.common;
-
-import io.openems.edge.common.channel.ChannelId;
-import io.openems.edge.common.channel.Doc;
-
-public class ChannelIdImpl implements ChannelId {
-	;
-
-	private final String name;
-	private final Doc doc;
-
-	public ChannelIdImpl(String name, Doc doc) {
-		this.name = name;
-		this.doc = doc;
-	}
-
-	@Override
-	public String name() {
-		return this.name;
-	}
-
-	@Override
-	public Doc doc() {
-		return this.doc;
-	}
-}
