+++ io.openems.edge.common/src/io/openems/edge/common/channel/ChannelId.java
@@ -102,4 +102,7 @@ public interface ChannelId {
 	 * @return the Channel-Doc
 	 */
 	Doc doc();
+	
+	public record ChannelIdImpl(String name, Doc doc) implements ChannelId {
+	}
 }
